% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouping.R
\name{read_sample_groups}
\alias{read_sample_groups}
\title{Read sample groups from a text file}
\usage{
read_sample_groups(file_path, sep = "\\t", header = TRUE)
}
\arguments{
\item{file_path}{Path to the group definition file}

\item{sep}{Separator used in the file (default: tab)}

\item{header}{Does the file contain a header line?}
}
\value{
A named list of sample groups
}
\description{
Read sample groups from a text file
}
\examples{
groups <- read_sample_groups("path/to/group.txt")
}
