% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_analysis_params}
\alias{set_analysis_params}
\title{Set analysis parameters}
\usage{
set_analysis_params(
  cell_threshold = 2000,
  pca_dims = 1:30,
  resolution_setting = 1,
  parent_marker = "_",
  min_pct = 0.25,
  logfc_threshold = 0.585
)
}
\arguments{
\item{cell_threshold}{Minimum cell count for reclustering}

\item{pca_dims}{Dimensions to use for PCA}

\item{resolution_setting}{Clustering resolution}

\item{parent_marker}{Separator for hierarchical clustering}

\item{min_pct}{Minimum percentage of cells expressing a gene}

\item{logfc_threshold}{Log fold change threshold for marker genes}
}
\value{
A list of analysis parameters
}
\description{
Set analysis parameters
}
\examples{
params <- set_analysis_params(cell_threshold = 1500, resolution_setting = 0.8)
}
